+++ src/onegov.activity/onegov/activity/collections/__init__.py
@@ -1,5 +1,11 @@
 from onegov.activity.collections.activity import ActivityCollection
 from onegov.activity.collections.booking import BookingCollection
 from onegov.activity.collections.occasion import OccasionCollection
+from onegov.activity.collections.period import PeriodCollection
 
-__all__ = ['ActivityCollection', 'BookingCollection', 'OccasionCollection']
+__all__ = [
+    'ActivityCollection',
+    'BookingCollection',
+    'OccasionCollection',
+    'PeriodCollection'
+]