+++ src/onegov.activity/onegov/activity/models/__init__.py
@@ -1,6 +1,14 @@
 from onegov.activity.models.activity import Activity, ACTIVITY_STATES
 from onegov.activity.models.booking import Booking
 from onegov.activity.models.occasion import Occasion, DAYS
+from onegov.activity.models.period import Period
 
 
-__all__ = ['Activity', 'Booking', 'Occasion', 'ACTIVITY_STATES', 'DAYS']
+__all__ = [
+    'Activity',
+    'Booking',
+    'Occasion',
+    'Period',
+    'ACTIVITY_STATES',
+    'DAYS'
+]