+++ src/onegov.activity/onegov/activity/upgrade.py
@@ -6,7 +6,7 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 from onegov.activity.models import Activity, Occasion
 from onegov.core.upgrade import upgrade_task
 from psycopg2.extras import NumericRange
-from sqlalchemy import Column, Text, Integer
+from sqlalchemy import Boolean, Column, Text, Integer
 from sqlalchemy.dialects.postgresql import ARRAY, INT4RANGE
 
 
@@ -131,3 +131,19 @@ def add_activity_ages_second_step(context):
         occasion.note = occasion.note and occasion.note.strip() or None
 
     context.session.flush()
+
+
+@upgrade_task('Rebuild models')
+def rebuild_models(context):
+    # delete the existing occasions/bookings table and start anew ->
+    # too many changes + we're at an early stage
+    context.operations.drop_table('bookings')
+    context.operations.drop_table('occasions')
+
+    context.operations.add_column(
+        'activities', Column('has_active_occasions', Boolean, default=False))
+
+
+@upgrade_task('Drop occasion state type')
+def drop_occasion_state_type(context):
+    context.operations.execute('DROP TYPE occasion_state')